refactor: Remove remaining imports from `src/utils.js`
