+++ src/config_loader.js
@@ -1,11 +1,9 @@
 "use strict";
-import {
-  isPureObject, isObject,
-  fetchTextFile, 
-  trim, 
-  deepMerge, 
-  removePrefix
-} from "./utils.js";
+import { isPureObject, isObject } from "./utils/type_check.js";
+import { removePrefix, trim } from "./utils/str_utils.js";
+import { deepMerge } from "./utils/deep_merge.js";
+import { fetchTextFile } from "./utils/file_load.js";
+
 import * as CNF_MOD from "./config.js";
